+++ Public/Restore-AGMLibSAPHANA.ps1
@@ -13,60 +13,46 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$mountapplianceid,[string]$imagename,[string]$imageid,[string]$targethostname,[string]$appname,[string]$recoverypoint,[string]$label,[string]$dbsid,[string]$userstorekey,[string]$mountpointperimage,[string]$mountmode,[string]$mapdiskstoallesxhosts,[string]$sltid,[string]$slpid,[string]$jsonprint,[switch][alias("g")]$guided,[switch][alias("m")]$monitor,[switch][alias("w")]$wait) 
+Function Restore-AGMLibSAPHANA ([string]$appid,[string]$targethostid,[string]$mountapplianceid,[string]$imagename,[string]$imageid,[string]$targethostname,[string]$appname,[string]$recoverypoint,[string]$label,[string]$dbsid,[string]$userstorekey,[string]$jsonprint,[switch][alias("g")]$guided,[switch]$preflight,[switch]$replacesource,[switch]$copyhdbstorekey) 
 {
     <#
     .SYNOPSIS
-    Mounts a SAP HANA Database Image
+    Restores a SAP HANA Database Image. Note that this command has different syntax requirements when issued to an AGM versus a Management Console.
 
     .EXAMPLE
-    New-AGMLibSAPHANAMount 
-    You will be prompted through a guided menu
+    Restore-AGMLibSAPHANA 
+    You will be prompted through a guided menu.  
 
     .EXAMPLE
-    New-AGMLibSAPHANAMount -appid 577110 -mountapplianceid 141767697828 -targethostid 483699 -dbsid "TGT" -userstorekey "ACTBACKUP" -mountpointperimage "/tgt"
-    Mounts an SAP HANA Database with new SID to the desired mountpoint.
+    Restore-AGMLibSAPHANA -appid 577110 -mountapplianceid 141767697828 -targethostid 483699 -dbsid "TGT" -userstorekey "ACTBACKUP"
+    Restores an SAP HANA Database with specified SID when using Google Cloud Backup and DR
+
+    .EXAMPLE
+    Restore-AGMLibSAPHANA -appid 577110 -userstorekey "ACTBACKUP"
+    Restores an SAP HANA Database when sent to Actifio Global Manager
 
     .DESCRIPTION
-    A function to mount SAP HANA Image
+    A function to restore a SAP HANA database
 
     * Image selection can be done three ways:
 
     1)  Run this command in guided mode to learn the available images and select one
     2)  Learn the imagename and specify that as part of the command with -imagename
-    3)  Learn the Appid and Cluster ID for the appliance that will mount the image and then use -appid and -mountapplianceid 
-    This will use the latest snapshot, StreamSnap or OnVault image on that appliance
+    3)  Learn the Appid and Cluster ID for the appliance that will restore the image and then use -appid and -mountapplianceid 
+    This will use the latest snapshot, StreamSnap or OnVault image on that appliance.  AGM does not need the Cluster ID
 
-    Note default values don't need to specified.  
+    * Restore host options (Google Cloud Backup and DR only)
+    -mountapplianceid  The appliance where the restore will be run.
+    -targethostname   The target host specified by name (targethostid can also be used)  
+    -targethostid   The target host specified by id (targethostname can also be used) 
 
-    * label
-    -label   Label for mount, recommended
+    *  required settings
+    -userstorekey xxxx   name of the HANA database user store key on the target server where a new SAP HANA Instance will get restored
 
-    * mount host options:
-    -targethostname   The target host specified by name.  Ideally use the next option
-    -targethostid   The target host specified by id
-
-    * mounted instance required settings
-    -userstorekey xxxx   name of the HANA database user store key on the target server where a new SAP HANA Instance will get created.
-    -mountpointperimage xxxx  path to the base directory where the configuration & database files for SAP HANA Instance on the target server are located.
-    
     * Other options
-
+     -label   Label for restore, recommended
     -recoverypoint  The point in time to roll forward to, in ISO8601 format like 2020-09-02 19:00:02
 
-    * Reprotection:
-
-    -sltid xxxx (short for Service Level Template ID) - if specified along with an slpid, will reprotect the mounted child app with the specified template and profile
-    -slpid yyyy (short for Service Level Profile ID) - if specified along with an sltid, will reprotect the mounted child app with the specified template and profile    
-
-    * VMware specific options
-    -mountmode    use either nfs, vrdm or prdm
-    -mapdiskstoallesxhosts   Either true to do this or false to not do this.  Default is false.  
-
-    * Monitoring options:
-
-    -wait     This will wait up to 2 minutes for the job to start, checking every 15 seconds to show you the job name
-    -monitor  Same as -wait but will also run Get-AGMLibFollowJobStatus to monitor the job to completion 
     #>
 
     # its pointless procededing without a connection.
@@ -140,8 +126,9 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
             $appname = $imagegrab.appname
             $appid = $imagegrab.application.id
             $apptype = $imagegrab.apptype      
-            $restorableobjects = $imagegrab.restorableobjects
             $mountapplianceid = $imagegrab.cluster.clusterid   
+            $dbsid = ($imagegrab.provisioningoptions | where-object {($_.key -eq "dbsid")}).value
+            if (!($userstorekey)) { $userstorekey = ($imagegrab.provisioningoptions | where-object {($_.key -eq "DBUSER")}).value }
         }
     }
 
@@ -161,8 +148,9 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
             $appname = $imagegrab.appname
             $appid = $imagegrab.application.id
             $apptype = $imagegrab.apptype      
-            $restorableobjects = $imagegrab.restorableobjects
             $mountapplianceid = $imagegrab.cluster.clusterid
+            $dbsid = ($imagegrab.provisioningoptions | where-object {($_.key -eq "dbsid")}).value
+            if (!($userstorekey)) { $userstorekey = ($imagegrab.provisioningoptions | where-object {($_.key -eq "DBUSER")}).value }
         }
     }
 
@@ -188,7 +176,7 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
         else
         {
             Clear-Host
-            write-host "Appliance selection menu - which Appliance will run this mount"
+            write-host "Appliance selection menu - which Appliance will run this restore"
             Write-host ""
             $i = 1
             foreach ($appliance in $appliancegrab)
@@ -200,7 +188,7 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
             {
                 Write-host ""
                 $listmax = $appliancegrab.name.count
-                [int]$appselection = Read-Host "Please select an Appliance to mount from (1-$listmax)"
+                [int]$appselection = Read-Host "Please select the Appliance that will run the restore (1-$listmax)"
                 if ($appselection -lt 1 -or $appselection -gt $listmax)
                 {
                     Write-Host -Object "Invalid selection. Please enter a number in range [1-$($listmax)]"
@@ -286,16 +274,6 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
         else {
             $hostgrab = Get-AGMHost -id $hostcheck.id
             $targethostid = $hostgrab.id
-            $vmtype = $hostgrab.vmtype
-            $transport = $hostgrab.transport
-            $diskpref = $hostgrab.diskpref
-            $vcenterid = $hostgrab.vcenterhost.id
-            #if the VM doesn't have a transport, then the vCenter must have one
-            if ( ($vmtype -eq "vmware") -and (!($transport)) )
-            {
-                $vcgrab = Get-AGMHost -filtervalue id=$vcenterid 
-                $transport = $vcgrab.transport
-            }
         }
     }
 
@@ -309,15 +287,6 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
         }
         $targethostid = $targethostid
         $targethostname=$hostgrab.hostname
-        $vmtype = $hostgrab.vmtype
-        $transport = $hostgrab.transport
-        $diskpref = $hostgrab.diskpref
-        $vcenterid = $hostgrab.vcenterhost.id
-        if ( ($vmtype -eq "vmware") -and (!($transport)) )
-        {
-            $vcgrab = Get-AGMHost -filtervalue id=$vcenterid 
-            $transport = $vcgrab.transport
-        }
     }
     
     # this if for guided menu
@@ -365,10 +334,12 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
                     $appname = $imagegrab.appname
                     $appid = $imagegrab.application.id
                     $apptype = $imagegrab.apptype      
-                    $restorableobjects = $imagegrab.restorableobjects | where-object {$_.systemdb -eq $false} 
                     $jobclass = $imagegrab.jobclass
                     $mountapplianceid = $imagegrab.cluster.clusterid
                     $mountappliancename = $imagegrab.cluster.name
+                    $sourcehostname = $imagegrab.host.hostname
+                    $sourcehostid = $imagegrab.host.id
+                    $sourcedbuser = ($imagegrab.provisioningoptions | where-object {($_.key -eq "DBUSER")}).value
                     write-host "Found one $jobclass image $imagename, consistency date $consistencydate on $mountappliancename"
                      Write-host ""
                  }
@@ -389,17 +360,20 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
                 $imagelist = $imagelist1  | Sort-Object consistencydate
                 if ($imagelist1.id.count -eq 1)
                 {
-                    $imagegrab = Get-AGMImage -id $($imagelist).id
+                    $imagegrab = Get-AGMImage -id $($imagelist1).id
                     $imagename = $imagegrab.backupname                
                     $consistencydate = $imagegrab.consistencydate
                     $endpit = $imagegrab.endpit
                     $appname = $imagegrab.appname
                     $appid = $imagegrab.application.id
                     $apptype = $imagegrab.apptype      
-                    $restorableobjects = $imagegrab.restorableobjects | where-object {$_.systemdb -eq $false} 
                     $jobclass = $imagegrab.jobclass
                     $mountapplianceid = $imagegrab.cluster.clusterid
                     $mountappliancename = $imagegrab.cluster.name
+                    $sourcehostname = $imagegrab.host.hostname
+                    $sourcehostid = $imagegrab.host.id
+
+                    $sourcedbuser = ($imagegrab.provisioningoptions | where-object {($_.key -eq "DBUSER")}).value
                     write-host "Found one $jobclass image $imagename, consistency date $consistencydate on $mountappliancename"
                 } 
                 else
@@ -439,9 +413,11 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
                     $appname = $imagegrab.appname
                     $appid = $imagegrab.application.id
                     $apptype = $imagegrab.apptype      
-                    $restorableobjects = $imagegrab.restorableobjects | where-object {$_.systemdb -eq $false}
                     $mountapplianceid = $imagegrab.cluster.clusterid
                     $mountappliancename = $imagegrab.cluster.name 
+                    $sourcehostname = $imagegrab.host.hostname
+                    $sourcehostid = $imagegrab.host.id
+                    $sourcedbuser = ($imagegrab.provisioningoptions | where-object {($_.key -eq "DBUSER")}).value
                 }
              }
         }
@@ -450,7 +426,7 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
         if ($endpit)
         {
             write-host ""
-            $recoverypoint = Read-Host "Roll forward time (hitting enter means no roll-forward)`: $consistencydate to $endpit"
+            $recoverypoint = Read-Host "Roll forward time (hitting enter to use all available logs)`: $consistencydate to $endpit"
             if ($recoverypoint)
             {
                 if ([datetime]$recoverypoint -lt $consistencydate)
@@ -464,226 +440,176 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
                     return
                 }
             }
+            else {
+                $recoverypoint = $endpit
+            }
         }
-    
+        
+
+
         if ( (!($targethostname)) -and (!($targethostid)))
         {
-            $hostgrab = Get-AGMHost -filtervalue "sourcecluster=$mountapplianceid" | sort-object name
-            if ($hostgrab -eq "" )
+            # target host
+            if ($AGMToken)
             {
-                Get-AGMErrorMessage -messagetoprint "Cannot find any Linux hosts"
-                return
-            }
-            Clear-Host
-            Write-Host "Target host selection menu"
-            $i = 1
-            $printarray = @()
-            foreach ($listedhost in $hostgrab)
-            { 
-                $printarray += [pscustomobject]@{
-                    id = $i
-                    hostname = $listedhost.name
-                    hostid = $listedhost.id
-                    vmtype = $listedhost.vmtype
-                }
-                $i++
-            }
-            #print the list
-            $printarray | Format-table 
-            write-host ""
-            While ($true) 
-            {
-                $listmax = $hostgrab.name.count
-                [int]$hostselection = Read-Host "Please select a host (1-$listmax)"
-                if ($hostselection -lt 1 -or $hostselection -gt $listmax)
+                write-host ""
+                $userselection = ""
+                Write-Host "Target host selection"
+                Write-Host "1`: Restore back to the source host (default)"
+                Write-Host "2`: Restore to a new target"
+                Write-Host ""
+                While ($true) 
                 {
-                    Write-Host -Object "Invalid selection. Please enter a number in range [1-$($listmax)]"
-                } 
-                else
-                {
-                    break
+                    [int]$targetselection = Read-Host "Please select from this list (1-2)"
+                    if ($targetselection -eq "") { $targetselection = 1 }
+                    if ($targetselection -lt 1 -or $targetselection -gt 2)
+                    {
+                        Write-Host -Object "Invalid selection. Please enter a number in range [1-2]"
+                    } 
+                    else
+                    {
+                        break
+                    }
                 }
+                write-host ""
+                $userstorekey = read-host "SAP HANA Target user store key or press enter to use $sourcedbuser"
             }
-            if ($hostgrab.name.count -eq 1)
+            else {
+                # we need to ask for SAB DB User Store-key
+                write-host ""
+                While ($true) 
+                {
+                    [string]$userstorekey = Read-Host "SAP DB User Store Key"
+                    if (!($userstorekey))
+                    {
+                        write-host "User key field cannot be blank" 
+                    } 
+                    else
+                    {
+                        break
+                    }
+                }
+                # we can only restore to source host
+                $targetselection = 1
+            }
+            if ($targetselection -eq 1)
             {
-                $targethostname =  $hostgrab.name
+                $targethostname = $sourcehostname
+                $targethostid = $sourcehostid
+            }
+            else 
+            {
+                $hostgrab = Get-AGMHost -filtervalue "sourcecluster=$mountapplianceid" | sort-object name
+                if ($hostgrab -eq "" )
+                {
+                    Get-AGMErrorMessage -messagetoprint "Cannot find any Linux hosts"
+                    return
+                }
+                Clear-Host
+                Write-Host "Target host selection menu"
+                $i = 1
+                $printarray = @()
+                foreach ($listedhost in $hostgrab)
+                { 
+                    $printarray += [pscustomobject]@{
+                        id = $i
+                        hostname = $listedhost.name
+                        hostid = $listedhost.id
+                        vmtype = $listedhost.vmtype
+                    }
+                    $i++
+                }
+                #print the list
+                $printarray | Format-table 
+                write-host ""
+                While ($true) 
+                {
+                    $listmax = $hostgrab.name.count
+                    [int]$hostselection = Read-Host "Please select a host (1-$listmax)"
+                    if ($hostselection -lt 1 -or $hostselection -gt $listmax)
+                    {
+                        Write-Host -Object "Invalid selection. Please enter a number in range [1-$($listmax)]"
+                    } 
+                    else
+                    {
+                        break
+                    }
+                }
+                if ($hostgrab.name.count -eq 1)
+                {
+                    $targethostname =  $hostgrab.name
+                    $targethostid = $hostgrab.id
+                } else {
+                    $targethostname =  $hostgrab.name[($hostselection - 1)]
+                    $targethostid = $hostgrab.id[($hostselection - 1)]
+                }
+                $hostgrab = Get-AGMHost -id $targethostid
                 $targethostid = $hostgrab.id
-            } else {
-                $targethostname =  $hostgrab.name[($hostselection - 1)]
-                $targethostid = $hostgrab.id[($hostselection - 1)]
-            }
-            $hostgrab = Get-AGMHost -id $targethostid
-            $targethostid = $hostgrab.id
-            $vmtype = $hostgrab.vmtype
-            $transport = $hostgrab.transport
-            $diskpref = $hostgrab.diskpref
-            $vcenterid = $hostgrab.vcenterhost.id
-            #if the VM doesn't have a transport, then the vCenter must have one
-            if ( ($vmtype -eq "vmware") -and (!($transport)) )
-            {
-                $vcgrab = Get-AGMHost -filtervalue id=$vcenterid 
-                $transport = $vcgrab.transport
+
+                 # target host
+                write-host ""
+                $replacehostselection = ""
+                Write-Host "REPLACE ORIGINAL APPLICATION IDENTITY"
+                Write-Host "1`: No (default)"
+                Write-Host "2`: Yes"
+                Write-Host ""
+                [int]$replacehostselection = Read-Host "Please select from this list (1-2)"
+
+                if ($replacehostselection -eq "") { $replacehostselection = 1 }
+                if ($replacehostselection -eq 1) { $replacesource = $false }
+                if ($replacehostselection -eq 2) { $replacesource = $true }
             }
         }
-        
-        # reprotection
-        Clear-Host
-        Write-Host "Reprotection"
-        Write-Host "1`: Don't manage new application (default)"
-        Write-Host "2`: Manage new application"
-        Write-Host ""
-        [int]$userselection = Read-Host "Please select from this list (1-2)"
-        if ($userselection -eq "") { $userselection = 1 }
-        if ($userselection -eq 2) 
-        {   
-            # slt selection
-            Clear-Host
-            Write-Host "SLT list"
-            $objectgrab = Get-AGMSLT | sort-object name
-            $i = 1
-            foreach
-            ($object in $objectgrab)
-                { Write-Host -Object "$i`:  $($object.name) ($($object.id))"
-                $i++
-            }
-            While ($true) 
-            {
-                Write-host ""
-                $listmax = $objectgrab.Length
-                [int]$objectselection = Read-Host "Please select from this list (1-$listmax)"
-                if ($objectselection -lt 1 -or $objectselection -gt $listmax)
-                {
-                    Write-Host -Object "Invalid selection. Please enter a number in range [1-$listmax]"
-                } 
-                else
-                {
-                    break
-                }
-            }
-            $sltid =  $objectgrab[($objectselection - 1)].id
+    
 
-            #slp selection
-            Clear-Host
-            Write-Host "SLP list"
-            $objectgrab = Get-AGMSLP -filtervalue clusterid=$mountapplianceid | sort-object name
-            $i = 1
-            foreach
-            ($object in $objectgrab)
-                { Write-Host -Object "$i`:  $($object.name) ($($object.id))"
-                $i++
-            }
-            While ($true) 
-            {
-                Write-host ""
-                $listmax = $objectgrab.Length
-                [int]$objectselection = Read-Host "Please select from this list (1-$listmax)"
-                if ($objectselection -lt 1 -or $objectselection -gt $listmax)
-                {
-                    Write-Host -Object "Invalid selection. Please enter a number in range [1-$listmax]"
-                } 
-                else
-                {
-                    break
-                }
-            }
-            $slpid =  $objectgrab[($objectselection - 1)].id
+        write-host ""
+        if ($userstorekey -eq "")
+        {
+            $userstorekey = $sourcedbuser
+        } 
+        if (!($AGMToken))
+        {
+            write-host ""
+            $userselection = ""
+            Write-Host "Copy HDB User Store Key to target Host"
+            Write-Host "1`: No (default)"
+            Write-Host "2`: Yes"
+            Write-Host ""
+            [int]$replacehostselection = Read-Host "Please select from this list (1-2)"
+
+            if ($replacehostselection -eq "") { $replacehostselection = 1 }
+            if ($replacehostselection -eq 1) { $copyhdbstorekey = $false }
+            if ($replacehostselection -eq 2) { $copyhdbstorekey = $true }
         }
 
-       
-        write-host ""
-        
-        While ($true) 
-        {
-            $dbsid = read-host "SAP HANA Target database SID"
-            if ($dbsid -eq "")
-            {
-                Write-Host -Object "The Target database SID cannot be blank"
-            } 
-            else
-            {
-                break
-            }
-        }
-        write-host ""
-        While ($true) 
-        {
-            $userstorekey = read-host "SAP HANA Target user store key"
-            if ($userstorekey -eq "")
-            {
-                Write-Host -Object "The Target user store key cannot be blank"
-            } 
-            else
-            {
-                break
-            }
-        }
-        write-host ""
-        While ($true) 
-        {
-            $mountpointperimage = read-host "SAP HANA Target filesystem mount point"
-            if ($mountpointperimage -eq "")
-            {
-                Write-Host -Object "The Target Server filesystem mount point cannot be blank"
-            } 
-            else
-            {
-                break
-            }
-        }
         Write-host ""
 
-
-        # if this is a VMTarget
-        if ($vmtype -eq "vmware")
-        {
-            if (($diskpref -eq "BLOCK") -and ($transport -ne "GUESTVMISCSI"))
-            {
-                Clear-Host
-                Write-Host "Mount mode" 
-                if ($transport -eq "NFS")
-                {
-                    $defaultmode = 3
-                    Write-Host "1`: vrdm"
-                    Write-Host "2`: prdm"
-                    Write-Host "3`: nfs(default)"
-                }
-                else 
-                {
-                    $defaultmode = 1
-                    Write-Host "1`: vrdm(default)"
-                    Write-Host "2`: prdm"
-                    Write-Host "3`: nfs"
-                }
-                Write-Host ""
-                [int]$userselection = Read-Host "Please select from this list (1-3)"
-                if ($userselection -eq "") { $userselection = $defaultmode }
-                if ($userselection -eq 1) {  $mountmode = "vrdm"  }
-                if ($userselection -eq 2) {  $mountmode = "prdm"  }
-                if ($userselection -eq 3) {  $mountmode = "nfs"  }
-        
-                # map to all ESX host 
-                Clear-Host
-                Write-Host "Map to all ESX Hosts"
-                Write-Host "1`: Do not map to all ESX Hosts(default)"
-                Write-Host "2`: Map to all ESX Hosts"
-                Write-Host ""
-                [int]$userselection = Read-Host "Please select from this list (1-2)"
-                if ($userselection -eq "") { $userselection = 1 }
-                if ($userselection -eq 1) {  $mapdiskstoallesxhosts = "false"  }
-                if ($userselection -eq 2) {  $mapdiskstoallesxhosts = "true"  }
-            }
-        }
-        #volume info section
-        #$logicalnamelist = $restorableobjects.volumeinfo | select-object logicalname,capacity,uniqueid | sort-object logicalname | Get-Unique -asstring
-        Clear-Host  
-        
-
+    
         # we are done
        Clear-Host  
-        Write-Host "Guided selection is complete. The values entered would result in the following command:"
+        Write-Host "Guided selection is complete. The values entered resulted in the following commands."
+        if ($AGMToken)
+        {
+            Write-host "You should run the pre-flight first (first command) to validate restore will run without error before running the actual restore (second command)."
+            Write-Host ""
+        
+            Write-Host -nonewline "Restore-AGMLibSAPHANA -appid $appid -mountapplianceid $mountapplianceid -imagename $imagename -targethostid $targethostid -userstorekey `"$userstorekey`""
+            if ($label)
+            {
+                Write-Host -nonewline " -label `"$label`""
+            }
+            if ($recoverypoint)
+            {
+                Write-Host -nonewline " -recoverypoint `"$recoverypoint`""
+            }
+            if ($replacesource -eq $true)
+            {
+                Write-Host -nonewline " -replacesource"
+            }
+            Write-Host -nonewline " -preflight"
+        }
         Write-Host ""
-       
-        Write-Host -nonewline "New-AGMLibSAPHANAMount -appid $appid -mountapplianceid $mountapplianceid -imagename $imagename -targethostid $targethostid -dbsid `"$dbsid`" -userstorekey `"$userstorekey`" -mountpointperimage `"$mountpointperimage`""
+        Write-Host -nonewline "Restore-AGMLibSAPHANA -appid $appid  -imagename $imagename -userstorekey `"$userstorekey`""
         if ($label)
         {
             Write-Host -nonewline " -label `"$label`""
@@ -692,54 +618,36 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
         {
             Write-Host -nonewline " -recoverypoint `"$recoverypoint`""
         }
-        if ($mountmode)
+        if ($replacesource -eq $true)
         {
-            Write-Host -nonewline " -mountmode $mountmode -mapdiskstoallesxhosts $mapdiskstoallesxhosts"
+            Write-Host -nonewline " -replacesource"
         }
-
-        if ($sltid)
+        if ($copyhdbstorekey -eq $true)
         {
-            Write-Host -nonewline " -sltid $sltid -slpid $slpid"
-        }        
+            Write-Host -nonewline " -copyhdbstorekey"
+        }
         Write-Host ""
-        Write-Host "1`: Run the command now (default)"
-        Write-Host "2`: Show the JSON used to run this command, but don't run it"
-        Write-Host "3`: Show the CSV to run this command with New-AGMLibSAPHANAMultiMount"
-        Write-Host "4`: Exit without running the command"
-        $userchoice = Read-Host "Please select from this list (1-3)"
-        if ($userchoice -eq 2)
-        {
-            $jsonprint = "yes"
-        }
-        if ($userchoice -eq 3)
-        {
-            write-host "appid,appname,mountapplianceid,imagename,targethostid,dbsid,userstorekey,mountpointperimage,label,recoverypoint,mountmode,mapdiskstoallesxhosts,sltid,slpid"
-            write-host -nonewline "$appid,`"$appname`",`"$mountapplianceid`",`"$imagename`",`"$targethostid`",`"$dbsid`",`"$userstorekey`",`"$mountpointperimage`",`"$label`",`"$recoverypoint`",`"$mountmode`",`"$mapdiskstoallesxhosts`",`"$sltid`",`"$slpid`""
-            write-host ""
-            return
-        }
-        if ($userchoice -eq 4)
-        {
-            return
-        }
-
-    }
-
-    if ($targethostid -eq "")
-    {
-        Get-AGMErrorMessage -messagetoprint "Cannot proceed without a targethostid or targethostname"
+        Read-Host "Please enter to exit"
         return
     }
+
+
 
     if (($appid) -and ($mountapplianceid) -and (!($imageid)))
     {
         # if we are not running guided mode but we have an appid without imageid, then lets get the latest image on the mountappliance ID
         $imagegrab = Get-AGMImage -filtervalue "appid=$appid&targetuds=$mountapplianceid&jobclass=snapshot&jobclass=StreamSnap&jobclass=OnVault" -sort "consistencydate:desc,jobclasscode:asc" -limit 1
-        if ($imagegrab.count -eq 1)
+        if ($imagegrab.id.count -eq 1)
         {   
-            $imageid = $imagegrab.id
-            $imagename = $imagegrab.backupname
-        }
+            $imagegrab1 = Get-AGMImage -id $($imagegrab).id
+            $imageid = $imagegrab1.id
+            $imagename = $imagegrab1.backupname
+            $targethostid = $imagegrab1.host.id
+            
+            $dbsid = ($imagegrab1.provisioningoptions | where-object {($_.key -eq "dbsid")}).value
+            if (!($userstorekey)) { $userstorekey = ($imagegrab1.provisioningoptions | where-object {($_.key -eq "DBUSER")}).value }
+            if ($imagegrab1.endpit) { $recoverypoint = $imagegrab1.endpit }
+        } 
         else 
         {
             Get-AGMErrorMessage -messagetoprint "Failed to fetch a snapshot, StreamSnap or OnVault Image for appid $appid on appliance with clusterID $mountapplianceid"
@@ -747,16 +655,12 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
         }
     }
 
-    if ( ($sltid) -and (!($slpid)) )
+    if (($targethostid -eq "") -and ($AGMToken))
     {
-        Get-AGMErrorMessage -messagetoprint "An sltid $sltid was specified without an slpid with -slpid yyyy. Please specify both"
+        Get-AGMErrorMessage -messagetoprint "Cannot proceed without a targethostid or targethostname"
         return
     }
-    if ( (!($sltid)) -and ($slpid) )
-    {
-        Get-AGMErrorMessage -messagetoprint "An slpid $slpid was specified without an sltid using -sltid xxxx. Please specify both"
-        return
-    }
+
     
     # learn about the image
     if (!($imagename)) 
@@ -776,11 +680,6 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
         $label = ""
     }
 
-    if (!($dbsid))
-    {
-        Get-AGMErrorMessage -messagetoprint "No database SID was specified"
-        return
-    }
 
     if (!($userstorekey))
     {
@@ -788,72 +687,8 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
         return
     }
 
-    if (!($mountpointperimage))
-    {
-        Get-AGMErrorMessage -messagetoprint "No Target directory mount point was specified"
-        return
-    }
-
-    if ($mountmode -eq "vrdm")
-    {
-        $physicalrdm = "0"
-        $rdmmode = "independentvirtual"
-    }
-    if ($mountmode -eq "prdm")
-    {
-        $physicalrdm = "1"
-        $rdmmode = "physical"
-    }
-    if ($mountmode -eq "nfs")
-    {
-        $physicalrdm = "2"
-        $rdmmode = "nfs"
-    }
-
-    if ($mapdiskstoallesxhosts)
-    {
-        if (($mapdiskstoallesxhosts -ne "true") -and  ($mapdiskstoallesxhosts -ne "false"))
-        {
-            Get-AGMErrorMessage -messagetoprint "The value of Map to all ESXi hosts of $mapdiskstoallesxhosts is not valid. Must be true or false"
-            return
-        }
-        $restoreoptions = @(
-            @{
-                name = 'mapdiskstoallesxhosts'
-                value = "$mapdiskstoallesxhosts"
-            }
-        )
-    }
-
-    if ($mountpointperimage)
-    {
-        if ($restoreoptions)
-        {
-            $imagemountpoint = @{
-                name = 'mountpointperimage'
-                value = "$mountpointperimage"
-            }
-            $restoreoptions = $restoreoptions + $imagemountpoint
-        }
-        else 
-        {
-            $restoreoptions = @(
-            @{
-                name = 'mountpointperimage'
-                value = "$mountpointperimage"
-            }
-        )
-        }
-    }
 
     $provisioningoptions = @()
-    if ($consistencygroupname)
-    {
-        $provisioningoptions = $provisioningoptions +[ordered]@{
-                name = 'ConsistencyGroupName'
-                value = $consistencygroupname
-            }
-    }
     $provisioningoptions = $provisioningoptions +[ordered]@{
         name = 'dbsid'
         value = $dbsid
@@ -862,98 +697,65 @@ Function New-AGMLibSAPHANAMount ([string]$appid,[string]$targethostid,[string]$m
         name = 'DBUSER'
         value = $userstorekey
     }
-    
-    if ($sltid)
-    {
-        $provisioningoptions= $provisioningoptions +[ordered]@{
-            name = 'reprotect'
-            value = "true"
-        }
-        $provisioningoptions= $provisioningoptions +[ordered]@{
-            name = 'slt'
-            value = $sltid
-        }
-        $provisioningoptions= $provisioningoptions +[ordered]@{
-            name = 'slp'
-            value = $slpid
-        }
-    }
-    $selectedobjects = @()
-    $selectedobjects = $selectedobjects + [ordered]@{
-        restorableobject = $appname
-    }
+    # this is for 10.0.x only
+    if (!($copyhdbstorekey)) { $copyhdbstorekey = $false}
+    # build body
     $body = [ordered]@{}
-    if ($rdmmode) { $body = $body + [ordered]@{ rdmmode = $rdmmode; }}
-    if ($physicalrdm) { $body = $body + [ordered]@{ physicalrdm = $physicalrdm; }}
     if ($label) { $body = $body + [ordered]@{ label = $label; }}
-    $body = $body + [ordered]@{
-        image = $imagename;
-        host = @{id=$targethostid};
-        hostclusterid = $mountapplianceid;
-        appaware = "true";
-        provisioningoptions = $provisioningoptions;
-        selectedobjects = $selectedobjects
-    }
-    if ($restoreoptions)
+    if ($AGMToken)
     {
-        $body = $body + @{ restoreoptions = $restoreoptions }
+        if ($replacesource -eq $true) { $body = $body + [ordered]@{ replacesource = $true; } } else { $body = $body + [ordered]@{ replacesource = $false; } }
+        $body = $body + [ordered]@{
+            host = @{id=$targethostid};
+            hostclusterid = $mountapplianceid;
+            provisioningoptions = $provisioningoptions;
+        }
+    }
+    else {
+        $restoreoptions = @()
+        $restoreoptions = $restoreoptions +[ordered]@{
+            name = 'copyhdbstorekey'
+            value = $copyhdbstorekey
+        }
+        $restoreoptions = $restoreoptions +[ordered]@{
+            name = 'restoreuserstorekey'
+            value = $userstorekey
+        }
+        
+        $body = $body + [ordered]@{
+            restoreoptions = $restoreoptions
+        }
     }
     if ($recoverytime)
     {
-        $body = $body + @{ recoverytime = [string]$recoverytime }
+        $body = $body +[ordered]@{ recoverytime = [string]$recoverytime }
     }
-    if ($restoreobjectmappings)
-    {
-        $body = $body + @{ restoreobjectmappings = $restoreobjectmappings }
-    }
+    
 
     $json = $body | ConvertTo-Json -depth 10
 
-    if ($monitor)
-    {
-        $wait = $true
-    }
 
     if ($jsonprint -eq "yes")
     {
         $compressedjson = $body | ConvertTo-Json -compress -depth 10
         Write-host "This is the final command:"
         Write-host ""
-        Write-host "Post-AGMAPIData -endpoint /backup/$imageid/mount -body `'$compressedjson`'"
+        Write-host "Post-AGMAPIData -endpoint /backup/$imageid/restore -body `'$compressedjson`'"
         return
     }
 
-    Post-AGMAPIData  -endpoint /backup/$imageid/mount -body $json
-
-    if ($wait)
+    if ($preflight)
     {
-        Start-Sleep -s 15
-        $i=1
-        while ($i -lt 9)
+        $preflighttest = Post-AGMAPIData  -endpoint /backup/$imageid/restorepreflight -body $json
+        if ($preflighttest.testlist)
         {
-            Clear-Host
-            write-host "Checking for a running job for appid $appid against targethostname $targethostname"
-            $jobgrab = Get-AGMJob -filtervalue "appid=$appid&jobclasscode=5&isscheduled=False&targethost=$targethostname" -sort queuedate:desc -limit 1 
-            if (!($jobgrab.jobname))
-            {
-                write-host "Job not running yet, will wait 15 seconds and check again. Check $i of 8"
-                Start-Sleep -s 15
-                $jobgrab = Get-AGMJob -filtervalue "appid=$appid&jobclasscode=5&isscheduled=False&targethost=$targethostname" -sort queuedate:desc -limit 1 
-                if (!($jobgrab.jobname))
-                {
-                    $i++
-                }
-            }
-            else
-            {   
-                $i=9
-                $jobgrab| select-object jobname,status,progress,queuedate,startdate,targethost
-                
-            }
+            $preflighttest.testlist
         }
-        if (($jobgrab.jobname) -and ($monitor))
-        {
-            Get-AGMLibFollowJobStatus $jobgrab.jobname
+        else {
+            $preflighttest
         }
+    }
+    else {
+        Post-AGMAPIData  -endpoint /backup/$imageid/restore -body $json
     }
 }