+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Change log
 
 Started Change log file as per release 0.0.0.54
+## AGMPowerLIB  (0.0.0.59)
+* Added Restore-AGMLibSAPHANA to run restores on SAP HANA
+* Added New-AGMLibSAPHANAMultiMount to do multi mount for SAP HANA
+* Updated Mount-AGMLibSAPHANA to add CSV output to feed New-AGMLibSAPHANAMultiMount
+* Minor fixes to New-AGMLibGCPInstance 
+* New-AGMLibImage will now look for DirectOnvault job
 
 ## AGMPowerLIB  (0.0.0.58)
 * Get-AGMLibPolicies - added options